refs #11 - Move code around
