+++ Interpreter/AST/Literals.cs
@@ -1,0 +1,165 @@
+using System.Text;
+
+namespace Interpreter;
+
+public abstract class Literal<T>: AST<T> {
+    protected T val;
+
+    public Literal(T val, string Type): base(Type){
+        this.val = val;
+    }
+
+    public T Val() {
+        return this.val;
+    }
+
+    public override dynamic Eval(Context ctx) {
+        return this.val;
+    }
+
+    // no need to override check
+}
+
+public class StringLiteral: Literal<string> {
+
+    public StringLiteral(string val): base(val, AST<object>.STRING) {}
+}
+
+public class IntLiteral : Literal<int> {
+
+    public IntLiteral(int val): base(val, AST<object>.INTEGER) {}
+}
+
+public class FloatLiteral : Literal<float> {
+
+    public FloatLiteral(float val): base(val, AST<object>.FLOAT) {}
+}
+
+public class BoolLiteral : Literal<bool> {
+
+    public BoolLiteral(bool val): base(val, AST<object>.BOOL) {}
+}
+
+// an IEnumerator but not an IEnumerator
+public class Terms: AST {
+    int start;
+    int end;
+    int index;
+    protected List<AST> ls;
+    // we don't know looking at the fields
+    // this is used to sum sequences
+    public bool IsInfinite;
+    Func<int, dynamic> term;
+
+    // lambda copy for the class
+    // we have to ways to define the class
+    // (we could separate the classes in Terms and Terms<int>) but it's "three strikes and you are out"
+    // not two
+    Func<Terms> copy;
+
+    public override string Type {
+        get {
+            if (this.start == this.end) {
+                return AST<object>.INTEGER;
+            }
+            return this.term(0).Type;
+        }
+    }
+
+    public Terms(List<AST> ls) {
+        this.index = -1;
+        this.start = 0;
+        this.end = ls.Count();
+        this.ls = ls;
+        this.term = (int i) => {return ls[i];};
+        this.IsInfinite = false;
+
+        // know which one to call
+        this.copy = () => new Terms(ls);
+    }
+
+    public Terms(int start, int? end): base(AST<int>.ToStr()) {
+        this.index = start - 1;
+        this.start = start;
+        if (end is null) {
+            this.end = Int32.MaxValue;
+            this.IsInfinite = true;
+        }
+        else {
+            this.end = (int) end;
+            this.IsInfinite = false;
+        }
+        this.term = (int index) => {return new IntLiteral(index);};
+
+        this.copy = () => new Terms(start, end);
+
+        this.ls = null;
+    }
+
+    public Terms Clone() {
+        return this.copy();
+    }
+
+    public override dynamic Eval(Context ctx) {
+        if (this.MoveNext()) {
+            return this.Current.Eval(ctx);
+        }
+        return null;
+    }
+
+    public bool MoveNext() {
+        index += 1;
+        // do not include range end [start, end)
+        if (this.index >= end) {
+            return false;
+        }
+
+        return true;
+    }
+
+    public AST Current {
+        get {
+            return this.term(this.index);
+        }
+    }
+
+    public IEnumerator<AST> GetEnumerator() {
+        for (int i = this.index + 1; i < this.end; i++) {
+            yield return this.term(i);
+        }
+    }
+
+    public override string ToString() {
+        StringBuilder str = new StringBuilder("{");
+        if (!this.IsInfinite) {
+            foreach(AST item in this) {
+                str.Append(item.ToString() + ", ");
+            }
+        }
+        else {
+            str.Append($"{this.start}...");
+        }
+        str.Append("}");
+        return str.ToString();
+    }
+
+}
+
+// block node... reimagined
+public class SequenceLiteral : Literal<Terms> {
+
+    public SequenceLiteral(Terms val): base(val, AST<object>.SEQUENCE) {}
+
+    public SequenceLiteral Clone() {
+        return new SequenceLiteral(this.val.Clone());
+    }
+
+    public override dynamic Eval(Context ctx) {
+        return this;
+    }
+
+    public override string ToString() {
+        return this.val.ToString();
+    }
+}
+
