+++ Interpreter/Parser.cs
@@ -7,778 +7,6 @@ using System.Reflection;
 
 namespace Interpreter;
 
-public class NameError : Exception {
-    public NameError() {}
-    public NameError(string message): base(message) {}
-    public NameError(string message, Exception inner): base(message, inner) {}
-}
-
-public class UnexpectedToken : Exception {
-    public UnexpectedToken() {}
-    public UnexpectedToken(string message): base(message) {}
-    public UnexpectedToken(string message, Exception inner): base(message, inner) {}
-}
-
-public class SyntaxError : Exception {
-    public SyntaxError() {}
-    public SyntaxError(string message): base(message) {}
-    public SyntaxError(string message, Exception inner): base(message, inner) {}
-}
-
-public class TypeError : Exception {
-    public TypeError() {}
-    public TypeError(string message): base(message) {}
-    public TypeError(string message, Exception inner): base(message, inner) {}
-}
-
-public class RuntimeError : Exception {
-    public RuntimeError() {}
-    public RuntimeError(string message): base(message) {}
-    public RuntimeError(string message, Exception inner): base(message, inner) {}
-}
-
-public static class BestGuess {
-    public static int Expression = 1;
-    public static int Declaration = 2;
-}
-
-public class Context : Dictionary<string, dynamic> {
-
-    public dynamic this[string key] {
-        get {
-            try {
-                return base[key];
-            }
-            catch(System.Collections.Generic.KeyNotFoundException) {
-                throw new NameError(key + " is not defined");
-            }
-        }
-        set {
-            base[key] = value;
-        }
-    }
-
-    public Context Clone() {
-        Context ret = new Context();
-        foreach(KeyValuePair<string, dynamic> entry in this) {
-            ret.Add(entry.Key, (dynamic) entry.Value);
-        }
-        return ret;
-    }
-}
-
-public class AST {
-
-    protected string type;
-
-    public virtual string Type {
-        get {return this.type;}
-        set {this.type = value;}
-    }
-
-    public AST(string Type="DYNAMIC") {
-        this.Type = Type;
-    }
-
-    public virtual dynamic Eval(Context ctx) {
-        return "";
-    }
-
-    public override string ToString() {
-       return this.Eval(new Context()).ToString();
-    }
-
-    // type check
-    public virtual Exception Check() {
-        return null;
-    }
-}
-
-public class AST<T>: AST {
-
-    public static string STRING = Tokens.STRING;
-    public static string INTEGER = Tokens.INTEGER;
-    public static string FLOAT = Tokens.FLOAT;
-    public static string BOOL = "BOOL";
-    public static string DYNAMIC = "DYNAMIC";
-    public static string SEQUENCE = "SEQUENCE";
-
-    public static HashSet<string> FloatOp = new HashSet<string>{FLOAT, INTEGER};
-    public static HashSet<string> BoolOp = new HashSet<string>{FLOAT, INTEGER, BOOL, STRING};
-    public static HashSet<string> StrOp = new HashSet<string>{STRING};
-    public static HashSet<string> SeqOp = new HashSet<string>{SEQUENCE};
-
-    public static Dictionary<string, HashSet<string>> Compatible = new Dictionary<string, HashSet<string>>{
-        {FLOAT, FloatOp},
-        {BOOL, FloatOp},
-        {STRING, StrOp},
-    };
-
-    public AST(string Type): base(Type) {}
-
-    public static Dictionary<string, Type> Types = new Dictionary<string, Type>{
-        {STRING, typeof(string)},
-        {INTEGER, typeof(int)},
-        {FLOAT, typeof(float)},
-        {BOOL, typeof(bool)},
-        // can't typeof(dynamic)
-        // this also means we don't care about the type (print)
-        {DYNAMIC, typeof(object)}
-        // not for SEQUENCE
-    };
-    public static Dictionary<Type, string> RevTypes = new Dictionary<Type, string>{
-        {typeof(string), STRING},
-        {typeof(int), INTEGER},
-        {typeof(float), FLOAT},
-        {typeof(bool), BOOL},
-        {typeof(object), DYNAMIC}
-        // no reverse search for typeof(object)
-    };
-
-    public static string ToStr() {
-        return RevTypes[typeof(T)];
-    }
-    public static string ToStr(Type type) {
-        return RevTypes[type];
-    }
-    public static Type ToType(string str) {
-        return Types[str];
-    }
-
-    public static Type InferType(AST<object> ast) {
-        return Types[ast.Type];
-    }
-}
-
-public abstract class Literal<T>: AST<T> {
-    protected T val;
-
-    public Literal(T val, string Type): base(Type){
-        this.val = val;
-    }
-
-    public T Val() {
-        return this.val;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        return this.val;
-    }
-
-    // no need to override check
-}
-
-public class StringLiteral: Literal<string> {
-
-    public StringLiteral(string val): base(val, AST<object>.STRING) {}
-}
-
-public class IntLiteral : Literal<int> {
-
-    public IntLiteral(int val): base(val, AST<object>.INTEGER) {}
-}
-
-public class FloatLiteral : Literal<float> {
-
-    public FloatLiteral(float val): base(val, AST<object>.FLOAT) {}
-}
-
-public class BoolLiteral : Literal<bool> {
-
-    public BoolLiteral(bool val): base(val, AST<object>.BOOL) {}
-}
-
-// an IEnumerator but not an IEnumerator
-public class Terms: AST {
-    int start;
-    int end;
-    int index;
-    protected List<AST> ls;
-    // we don't know looking at the fields
-    // this is used to sum sequences
-    public bool IsInfinite;
-    Func<int, dynamic> term;
-
-    // lambda copy for the class
-    // we have to ways to define the class
-    // (we could separate the classes in Terms and Terms<int>) but it's "three strikes and you are out"
-    // not two
-    Func<Terms> copy;
-
-    public override string Type {
-        get {
-            if (this.start == this.end) {
-                return AST<object>.INTEGER;
-            }
-            return this.term(0).Type;
-        }
-    }
-
-    public Terms(List<AST> ls) {
-        this.index = -1;
-        this.start = 0;
-        this.end = ls.Count();
-        this.ls = ls;
-        this.term = (int i) => {return ls[i];};
-        this.IsInfinite = false;
-
-        // know which one to call
-        this.copy = () => new Terms(ls);
-    }
-
-    public Terms(int start, int? end): base(AST<int>.ToStr()) {
-        this.index = start - 1;
-        this.start = start;
-        if (end is null) {
-            this.end = Int32.MaxValue;
-            this.IsInfinite = true;
-        }
-        else {
-            this.end = (int) end;
-            this.IsInfinite = false;
-        }
-        this.term = (int index) => {return new IntLiteral(index);};
-
-        this.copy = () => new Terms(start, end);
-
-        this.ls = null;
-    }
-
-    public Terms Clone() {
-        return this.copy();
-    }
-
-    public override dynamic Eval(Context ctx) {
-        if (this.MoveNext()) {
-            return this.Current.Eval(ctx);
-        }
-        return null;
-    }
-
-    public bool MoveNext() {
-        index += 1;
-        // do not include range end [start, end)
-        if (this.index >= end) {
-            return false;
-        }
-
-        return true;
-    }
-
-    public AST Current {
-        get {
-            return this.term(this.index);
-        }
-    }
-
-    public IEnumerator<AST> GetEnumerator() {
-        for (int i = this.index + 1; i < this.end; i++) {
-            yield return this.term(i);
-        }
-    }
-
-    public override string ToString() {
-        StringBuilder str = new StringBuilder("{");
-        if (!this.IsInfinite) {
-            foreach(AST item in this) {
-                str.Append(item.ToString() + ", ");
-            }
-        }
-        else {
-            str.Append($"{this.start}...");
-        }
-        str.Append("}");
-        return str.ToString();
-    }
-
-}
-
-// block node... reimagined
-public class SequenceLiteral : Literal<Terms> {
-
-    public SequenceLiteral(Terms val): base(val, AST<object>.SEQUENCE) {}
-
-    public SequenceLiteral Clone() {
-        return new SequenceLiteral(this.val.Clone());
-    }
-
-    public override dynamic Eval(Context ctx) {
-        return this;
-    }
-
-    public override string ToString() {
-        return this.val.ToString();
-    }
-}
-
-// <In, Out>
-public abstract class BinaryOperation<T, R>: AST<R> {
-    public AST left;
-    public AST right;
-
-    public BinaryOperation(AST left, AST right): base(AST<R>.ToStr(typeof(R))) {
-        this.left = left;
-        this.right = right;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        // ValidateArgs -> See Type and check if they are valid before eval
-        try {
-            return this.Operation(this.left.Eval(ctx), this.right.Eval(ctx));
-        }
-        catch (Microsoft.CSharp.RuntimeBinder.RuntimeBinderException) {
-            string msg = $"Unsupported operand type(s) for {this.GetType().Name.ToLower()}: {this.left.Type} and {this.right.Type}";
-            throw new RuntimeError(msg);
-        }
-    }
-
-    public override Exception Check() {
-        bool dynamic_expr = this.right.Type == AST<object>.DYNAMIC || this.left.Type == AST<object>.DYNAMIC;
-        bool right_type = AST<object>.Compatible[this.Type].Contains(this.left.Type) && AST<object>.Compatible[this.Type].Contains(this.right.Type);
-
-        if (dynamic_expr || right_type) {
-            return null;
-        }
-
-        string msg = $"Unsupported operand type(s) for {this.GetType().Name.ToLower()}: {left.GetType()} and {right.GetType()} (expected {AST<R>.ToStr(typeof(R))})";
-        return new TypeError(msg);
-    }
-
-    public abstract R Operation(T a, T b);
-}
-
-public class Sum : BinaryOperation<float, float> {
-
-    public Sum(AST left, AST right): base(left, right) {}
-   
-    public override float Operation(float a, float b) {
-        return a + b;
-    }
-}
-
-public class Substraction : BinaryOperation<float, float> {
-
-    public Substraction(AST left, AST right): base(left, right) {}
-   
-    public override float Operation(float a, float b) {
-        return a + -b;
-    }
-}
-public class Division : BinaryOperation<float, float> {
-
-    public Division(AST left, AST right): base(left, right) {}
-   
-    public override float Operation(float a, float b) {
-        return a / b;
-    }
-}
-public class Mult : BinaryOperation<float, float> {
-   
-    public Mult(AST left, AST right): base(left, right) {}
-
-    public override float Operation(float a, float b) {
-        return a * b;
-    }
-}
-public class Modulo : BinaryOperation<float, float> {
-   
-    public Modulo(AST left, AST right): base(left, right) {}
-
-    public override float Operation(float a, float b) {
-        return a % b;
-    }
-}
-public class Exp : BinaryOperation<float, float> {
-
-    public Exp(AST left, AST right): base(left, right) {}
-   
-    public override float Operation(float a, float b) {
-        return (float) System.Math.Pow(a, b);
-    }
-}
-public class Equals : BinaryOperation<float, bool> {
-
-    public Equals(AST left, AST right): base(left, right) {}
-   
-    public override bool Operation(float a, float b) {
-        return a == b;
-    }
-}
-public class Higher : BinaryOperation<float, bool> {
-
-    public Higher(AST left, AST right): base(left, right) {}
-   
-    public override bool Operation(float a, float b) {
-        return a > b;
-    }
-}
-public class HigherEqual : BinaryOperation<float, bool>
-{
-
-    public HigherEqual(AST left, AST right) : base(left, right) { }
-
-    public override bool Operation(float a, float b)
-    {
-        return a >= b;
-    }
-}
-public class Lower : BinaryOperation<float, bool> {
-
-    public Lower(AST left, AST right): base(left, right) {}
-   
-    public override bool Operation(float a, float b) {
-        return a < b;
-    }
-}
-public class LowerEqual : BinaryOperation<float, bool>
-{
-
-    public LowerEqual(AST left, AST right) : base(left, right) { }
-
-    public override bool Operation(float a, float b)
-    {
-        return a <= b;
-    }
-}
-
-//
-public abstract class UnaryOperation<T, R> : AST<R> {
-
-    public AST block;
-
-    public override string Type {
-        get {
-            return this.block.Type;
-        }
-    }
-
-    public UnaryOperation(AST block): base(AST<R>.ToStr()) {
-        this.block = block;
-    }
-    
-    public override dynamic Eval(Context ctx) {
-        return this.Operation(this.block.Eval(ctx));
-    }
-
-    public abstract R Operation(T a);
-
-    public void Check() {
-        bool dynamic_expr = this.block.Type == AST<object>.DYNAMIC;
-        bool right_type = this.block.Type == AST<R>.ToStr(typeof(R));
-
-        if ((dynamic_expr || right_type)) {
-            return;
-        }
-
-        string msg = $"Unsupported operand type(s) for {this.GetType().Name.ToLower()}: {this.block.Type}";
-        throw new RuntimeError(msg);
-    }
-}
-
-public class ChangeSign : UnaryOperation<float, float> {
-
-    public ChangeSign(AST block): base(block) {}
-    
-    public override float Operation(float a) {
-        return -a;
-    }
-}
-
-public class Negate : UnaryOperation<float, bool> {
-
-    public Negate(AST block): base(block) {}
-
-    // c# gets confused with the dynamic type for some reason
-    public override dynamic Eval(Context ctx) {
-        return this.Operation(this.block.Eval(ctx));
-    }
-
-    public override bool Operation(float a) {
-        return this.Operation(Convert.ToBoolean(a));
-    }
-
-    public bool Operation(bool a) {
-        return !a;
-    }
-}
-
-// these AST's types depend of something so Type is a property
-public class VariableDeclaration: AST {
-
-    public string name;
-    public AST expression;
-
-    public override string Type {
-        get {
-            return this.expression.Type;
-        }
-    }
-
-    public VariableDeclaration(string name, AST expression) {
-        this.name = name;
-        this.expression = expression;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        ctx[this.name] = this.expression.Eval(ctx);
-        return null;
-    }
-
-    public override string ToString() {
-        if (this.expression is null) {
-            return "<(Variable) [name: " + this.name + ", value: undefined]>";
-        }
-        return "<(Variable) [name: " + this.name + ", value: " + this.expression.ToString() + "]>";
-    }
-}
-
-// this depends of the type of the declaration
-public class Variable : AST {
-
-    public string name;
-    // XXX notice this is not VariableDeclaration<dynamic>
-    // we only care about VariableDeclaration.expression so this is what is here
-    public VariableDeclaration declaration;
-
-    public override string Type{
-        get {
-            if (declaration is null) {
-                return AST<object>.DYNAMIC;
-            }
-            return declaration.Type;
-        }
-    }
-
-    public Variable(string name) {
-        this.name = name;
-        // we don't know beforehand
-        this.declaration = null;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        // should be a literal
-        return ctx[this.name];
-    }
-
-    public override string ToString() {
-        return "<(Variable) [name: " + this.name + "]>";
-    }
-}
-
-// always dynamic; we just don't know
-// that said we can change it (in case of sequences it's a must)
-public class BlockNode: AST {
-    // Contains a set of evaluables
-    // only the last block is returned to the interpreter
-
-    public List<AST> blocks;
-
-    public BlockNode(List<AST> blocks) {
-        this.blocks = blocks;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        int counter = blocks.Count();
-        foreach(AST block in this.blocks) {
-            if (counter == 1) {
-                var bl = block.Eval(ctx);
-                return bl;
-            }
-            block.Eval(ctx);
-            counter -= 1;
-        }
-        return null;
-    }
-
-    public override Exception Check() {
-        foreach(AST item in this.blocks) {
-            Exception? exc = item.Check();
-            if (!(exc is null)) {
-                return exc;
-            }
-        }
-        return null;
-    }
-
-    public override string ToString() {
-        return this.blocks.ToString();
-    }
-}
-
-public class FunctionDeclaration: AST {
-
-    public string name;
-    // we don't konw the types of each argument
-    public BlockNode args;
-    public AST body;
-
-    public override string Type{
-        get {
-            return this.body.Type;
-        }
-    }
-
-    public FunctionDeclaration(string name, BlockNode args, AST body) {
-        this.name = name;
-        this.args = args;
-        this.body = body;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        ctx[this.name] = this;
-        return null;
-    }
-
-    public override string ToString() {
-        return $"<(FunctionDeclaration) [name: {this.name}, args: {this.args}, body: {this.body} type: {this.Type}]>";
-     }
-
-
-    public override Exception Check() {
-        return this.body.Check();
-    }
-}
-
-// same as with Variable
-public class Function : AST {
-    public string name;
-    public BlockNode args;
-
-    public FunctionDeclaration declaration;
-
-    public override string Type {
-        get {
-            if (this.declaration is null) {
-                return AST<object>.DYNAMIC;
-            }
-            return this.declaration.Type;
-        }
-    }
-
-    public Function(string name, BlockNode args) {
-        this.name = name;
-        this.args = args;
-
-        this.declaration = null;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        FunctionDeclaration fun_decl = (FunctionDeclaration) ctx[this.name];
-        BlockNode fun_args = fun_decl.args;
-
-        // know if we have to infer the type (or if we already inferred it from the declaration)
-        bool is_dynamic = this.Type == AST<object>.DYNAMIC;
-        
-        Context fun_ctx = ctx.Clone();
-        for (int i = 0; i < fun_args.blocks.Count(); i++) {
-            Variable variable = (Variable) fun_args.blocks[i];
-            try {
-                AST expression = this.args.blocks[i];
-                fun_ctx[variable.name] = expression.Eval(ctx);
-            }
-            catch (Exception) {
-                throw new RuntimeError($"Too many/few arguments for {this.name}");
-            }
-        }
-
-        // allow recursivity (explicitly)
-        //fun_ctx[this.name] = fun_decl;
-
-        return fun_decl.body.Eval(fun_ctx);
-   }
-
-
-    public override string ToString() {
-        return $"<(Function) [name: {this.name}, args: {this.args}, type: {this.Type}]>";
-    }
-
-    // TODO we can pass parameters to check or do something to check the declaration
-    // based on the variables passed
-    // public override Exception Check() {
-}
-
-public class Lambda : AST {
-    // let-in expression
-
-    public BlockNode variables;
-    public AST body;
-
-    public override string Type {
-        get {
-            return body.Type;
-        }
-    }
-
-    public Lambda(BlockNode variables, AST body) {
-        this.variables = variables;
-        this.body = body;
-    }
-
-    public override dynamic Eval(Context ctx) {
-        Context local_ctx = ctx.Clone();
-
-        // https://github.com/matcom/programming/tree/main/projects/hulk#variables
-        // "( ... ) Fuera de una expresión let-in las variables dejan de existir. ( ... )"
-        // thus, we declare variables inside the scope of the lambda
-        this.variables.Eval(local_ctx);
-
-        var res = this.body.Eval(local_ctx);
-
-        return res;
-    }
-
-
-    public override Exception Check() {
-        return this.body.Check();
-    }
-}
-
-public class Conditional : AST {
-    // we convert it to bool
-    AST hypothesis;
-    AST thesis;
-    AST antithesis;
-
-    public override string Type {
-        get {
-            return this.thesis.Type;
-        }
-    }
-
-    public Conditional(
-        AST hypothesis,
-        AST thesis,
-        AST antithesis
-    ) {
-        this.hypothesis = hypothesis;
-        this.thesis = thesis;
-        this.antithesis = antithesis;
-   }
-
-    public override dynamic Eval(Context ctx) {
-        bool res = Convert.ToBoolean(this.hypothesis.Eval(ctx));
-        if (res) {
-            return this.thesis.Eval(ctx);
-        }
-        return this.antithesis.Eval(ctx);
-    }
-
-    public override Exception Check() {
-        // TODO thesis and antithesis have the same type (or dynamic etc etc)
-        // NOTE we need not to check the type of the hypothesis--everything evals to either True or False
-        // almost the same as binaryops.Check()
-
-        bool dynamic_expr = this.thesis.Type == AST<object>.DYNAMIC || this.antithesis.Type == AST<object>.DYNAMIC;
-        bool right_type = AST<object>.Compatible[this.thesis.Type].Contains(this.antithesis.Type);
-
-        if (dynamic_expr || right_type) {
-            return null;
-        }
-
-        string msg = $"Type mismatch inside Conditional statement : {this.thesis.Type} != {this.antithesis.Type}";
-        return new TypeError(msg);
-    }
-}
 
 public class Parser {
     // FunctionMap -> Dict with declared functions (they are global)
@@ -1374,136 +602,5 @@ public class Parser {
         }
 
         return new BlockNode(nodes);
-    }
-}
-
-// builtins
-//
-// class [Fn]BlockNode -> evaluable body
-// class [Fn] {
-//     public [Fn] {
-//         "[fn]",
-//         args,
-//         [Fn]BlockNode()
-//     }
-// }
-class PrintBlockNode : UnaryOperation<object, object> {
-    public PrintBlockNode(AST argument) : base(argument) {}
-
-    public override object Operation(object arg) {
-        Console.WriteLine(arg);
-        return null;
-    }
-}
-
-class Print : FunctionDeclaration {
-
-    public Print() : base(
-        "print",
-        new BlockNode(new List<AST>{new Variable("val")}),
-        new PrintBlockNode(new Variable("val"))
-    ) {}
-}
-
-class CosBlockNode : UnaryOperation<float, float> {
-    public CosBlockNode(AST argument) : base(argument) {}
-
-    public override float Operation(float arg) {
-        return (float) Math.Cos(arg);
-    }
-}
-
-class Cos : FunctionDeclaration {
-
-    public Cos() : base(
-        "cos",
-        new BlockNode(new List<AST>{new Variable("val")}),
-        new CosBlockNode(new Variable("val"))
-    ) {}
-}
-
-class SinBlockNode : UnaryOperation<float, float> {
-    public SinBlockNode(AST argument) : base(argument) {}
-
-    public override float Operation(float arg) {
-        return (float) Math.Sin(arg);
-    }
-}
-
-class Sin : FunctionDeclaration {
-
-    public Sin() : base(
-        "sin",
-        new BlockNode(new List<AST>{new Variable("val")}),
-        new SinBlockNode(new Variable("val"))
-    ) {}
-}
-
-class LogBlockNode : UnaryOperation<float, float> {
-    public LogBlockNode(AST argument) : base(argument) {}
-
-    public override float Operation(float arg) {
-        return (float) Math.Log(arg);
-    }
-}
-
-class Log : FunctionDeclaration {
-
-    public Log() : base(
-        "log",
-        new BlockNode(new List<AST>{new Variable("val")}),
-        new LogBlockNode(new Variable("val"))
-    ) {}
-}
-
-// Adding new tokens, you say?
-// nyahahahaha
-class True : VariableDeclaration {
-
-    public True() : base("True", new BoolLiteral(true)) {}
-}
-
-class False : VariableDeclaration {
-
-    public False() : base("False", new BoolLiteral(false)) {}
-}
-
-public class Interpreter {
-    public Context GLOBAL_SCOPE = new Context();
-    public Parser parser;
-    public AST _tree;
-    public BlockNode BUILTINS = new BlockNode(
-        new List<AST>{
-            new Print(), new Cos(), new Sin(), new Log(), new True(), new False()
-        }
-    );
-
-    public Interpreter(Parser parser) {
-        this.parser = parser;
-        this._tree = null;
-
-        BUILTINS.Eval(this.GLOBAL_SCOPE);
-    }
-
-    public AST tree {
-        get {
-            if (this._tree is null) {
-                this._tree = this.parser.Parse();
-            }
-            return this._tree;
-        }
-        set {}
-    }
-
-    public dynamic Interpret() {
-        AST tree = this.parser.Parse();
-        if (tree is null) {
-            return "";
-        }
-        var eval = tree.Eval(this.GLOBAL_SCOPE);
-        if (eval is null) {
-            return "";
-        }
-        return eval;
     }
 }
